requirements.txt
aiobotocore==2.21.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiolimiter==1.2.1
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
botocore==1.37.1
cachetools==5.5.2
certifi==2025.1.31
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
multidict==6.3.2
propcache==0.3.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
pytube==15.0.0
pytubefix==8.12.3
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.13.1
tzlocal==5.3.1
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
